+++ src/app/types/notion/NotionMeetingNotes.ts
@@ -1,0 +1,7 @@
+import { Collection, ObjectId } from 'mongodb';
+
+export interface NotionMeetingNotes extends Collection {
+	_id: ObjectId,
+	guild: string,
+	databaseId: string,
+}